refactor: cleanup round 3
